refactor: move `domEvent` to `dom`
